handle partial work + CI failures
