+++ crates/dispatcher/src/monitor.rs
@@ -46,6 +46,9 @@ async fn run_scans(state: &Arc<DispatcherState>) {
     if let Err(e) = scan_archival(state).await {
         warn!("archival scan failed: {e}");
     }
+    if let Err(e) = scan_ci_status(state).await {
+        warn!("CI status scan failed: {e}");
+    }
     // Clear stale rate limit state so it doesn't permanently block dispatching
     scan_rate_limit_staleness(state).await;
 }
@@ -192,6 +195,134 @@ async fn scan_archival(state: &Arc<DispatcherState>) -> DispatcherResult<()> {
         // Set TTL on jobs KV entry (NATS KV doesn't support per-key TTL post-creation,
         // so we just leave it — the job will be queryable via API until the bucket ages it out)
         debug!(key, "job archived");
+    }
+
+    Ok(())
+}
+
+async fn scan_ci_status(state: &Arc<DispatcherState>) -> DispatcherResult<()> {
+    let now = Utc::now();
+
+    // Find InReview jobs with ci_status == Pending and a PR URL
+    let candidates: Vec<(String, String, String)> = state
+        .jobs
+        .iter()
+        .filter(|e| {
+            let job = e.value();
+            job.state == JobState::InReview
+                && job.ci_status == Some(CiStatus::Pending)
+                && job.pr_url.is_some()
+        })
+        .map(|e| {
+            let job = e.value();
+            (
+                e.key().clone(),
+                job.pr_url.clone().unwrap(),
+                job.repo.clone(),
+            )
+        })
+        .collect();
+
+    if candidates.is_empty() {
+        return Ok(());
+    }
+
+    let (forgejo_url, forgejo_token) =
+        match (&state.config.forgejo_url, &state.config.forgejo_token) {
+            (Some(url), Some(token)) => (url.clone(), token.clone()),
+            _ => return Ok(()), // No Forgejo config, skip CI check
+        };
+
+    let forgejo = chuggernaut_forgejo_api::ForgejoClient::new(&forgejo_url, &forgejo_token);
+
+    for (job_key, pr_url, repo) in candidates {
+        let job = match state.jobs.get(&job_key) {
+            Some(j) => j.clone(),
+            None => continue,
+        };
+
+        // Check CI poll timeout
+        if let Some(check_since) = job.ci_check_since {
+            let elapsed = (now - check_since).num_seconds() as u64;
+            if elapsed > state.config.ci_poll_timeout_secs {
+                info!(
+                    job_key,
+                    elapsed, "CI poll timeout, treating as success (no CI configured?)"
+                );
+                let event = CiCheckEvent {
+                    job_key: job_key.clone(),
+                    pr_url: pr_url.clone(),
+                    ci_status: CiStatus::Success,
+                    detected_at: now,
+                };
+                publish_monitor_event(state, &subjects::MONITOR_CI_CHECK, &event).await;
+                continue;
+            }
+        }
+
+        // Parse owner/repo from job.repo (format: "owner/repo")
+        let parts: Vec<&str> = repo.splitn(2, '/').collect();
+        if parts.len() != 2 {
+            continue;
+        }
+        let (owner, repo_name) = (parts[0], parts[1]);
+
+        // Get PR index from URL
+        let pr_index = match parse_pr_url_index(&pr_url) {
+            Some(idx) => idx,
+            None => continue,
+        };
+
+        // Get PR to find head SHA
+        let pr = match forgejo.get_pull_request(owner, repo_name, pr_index).await {
+            Ok(pr) => pr,
+            Err(e) => {
+                debug!(job_key, error = %e, "failed to get PR for CI check");
+                continue;
+            }
+        };
+
+        // Get combined commit status for the head SHA
+        match forgejo
+            .get_combined_status(owner, repo_name, &pr.head.sha)
+            .await
+        {
+            Ok(status) => {
+                // No CI configured (no statuses at all) — treat as success
+                if status.total_count == 0 {
+                    info!(job_key, "no CI statuses found, treating as success");
+                    let event = CiCheckEvent {
+                        job_key: job_key.clone(),
+                        pr_url: pr_url.clone(),
+                        ci_status: CiStatus::Success,
+                        detected_at: now,
+                    };
+                    publish_monitor_event(state, &subjects::MONITOR_CI_CHECK, &event).await;
+                    continue;
+                }
+
+                let ci_status = match status.state.as_str() {
+                    "success" => CiStatus::Success,
+                    "failure" => CiStatus::Failure,
+                    "error" => CiStatus::Error,
+                    _ => CiStatus::Pending, // "pending" or ""
+                };
+
+                if ci_status != CiStatus::Pending {
+                    let event = CiCheckEvent {
+                        job_key: job_key.clone(),
+                        pr_url: pr_url.clone(),
+                        ci_status,
+                        detected_at: now,
+                    };
+                    publish_monitor_event(state, &subjects::MONITOR_CI_CHECK, &event).await;
+                }
+                // else: still pending, will check again next scan
+            }
+            Err(e) => {
+                debug!(job_key, error = %e, "CI status check failed, will retry next scan");
+            }
+        }
     }
 
     Ok(())
